refactor(ts): migrar todas las pages/ a TypeScript
- Dashboard, LoginPage, GestionProductosPagina, SobreMiPage
- VendedoresPendientes, VentasPagina, VentasPendientesPagina
- Eliminar prop isMobile muerta en Dashboard (AppLayout no la usa)
